+++ scripts/2.79/kinect_mocap.py
@@ -23,9 +23,8 @@ bl_info = {
     "name": "Kinect Motion Capture plugin",
     "description": "Motion capture using MS Kinect v2",
     "author": "Morgane Dufresne",
-    "version": (1, 3),
+    "version": (1, 4),
     "blender": (2, 79, 0),
-    "warning": "You need a MS Kinect v2 sensor (XBox One)",
     "support": "COMMUNITY",
     "category": "Animation"
 }
@@ -57,6 +56,32 @@ def validateTarget(self, context):
             self.value = ""
     return None
 
+KBonesEnum = [("Head", "Head", "Head"),
+    ("Neck", "Neck", "Neck"),
+    ("Spine1", "Spine1", "Spine1"),
+    ("Spine0", "Spine0", "Spine0"),
+    ("LeftShoulder", "LeftShoulder", "LeftShoulder"),
+    ("LeftUpperArm", "LeftUpperArm", "LeftUpperArm"),
+    ("LeftLowerArm", "LeftLowerArm", "LeftLowerArm"),
+    ("LeftHand", "LeftHand", "LeftHand"),
+    ("RightShoulder", "RightShoulder", "RightShoulder"),
+    ("RightUpperArm", "RightUpperArm", "RightUpperArm"),
+    ("RightLowerArm", "RightLowerArm", "RightLowerArm"),
+    ("RightHand", "RightHand", "RightHand"),
+    ("LeftUpperLeg", "LeftUpperLeg", "LeftUpperLeg"),
+    ("LeftLowerLeg", "LeftLowerLeg", "LeftLowerLeg"),
+    ("LeftFoot", "LeftFoot", "LeftFoot"),
+    ("RightUpperLeg", "RightUpperLeg", "RightUpperLeg"),
+    ("RightLowerLeg", "RightLowerLeg", "RightLowerLeg"),
+    ("RightFoot", "RightFoot", "RightFoot")
+]
+
+KalmanStrengthEnum = [("Strong", "Strong", "Strong denoising (ideal for slow movement)"),
+    ("Normal", "Normal", "Normal denoising (good balance between speed and precision)"),
+    ("Low", "Low", "Low denoising (for fast movement, less precise)"),
+    ("VeryLow", "Very low", "Very low denoising (only for very fast movement, almost no noise reduction")
+]
+
 class KmcTarget(bpy.types.PropertyGroup):
     name = bpy.props.StringProperty(name="KBone")
     value = bpy.props.StringProperty(name="TBone", update=validateTarget)
@@ -72,7 +97,8 @@ class KmcProperties(bpy.types.PropertyGroup):
     lockHeight = bpy.props.BoolProperty(name="height", description="ignore vertical movement", default=True)
     lockwidth = bpy.props.BoolProperty(name="width", description="ignore lateral movement", default=True)
     lockDepth = bpy.props.BoolProperty(name="depth", description="ignore depth movement", default=True)
-
+    rootBone = bpy.props.EnumProperty(name="root bone", items=KBonesEnum, default="Spine0", description="Kinect identifier of the bone that is used as root of the skeleton")
+    kalmanStrength = bpy.props.EnumProperty(name="Denoising", items=KalmanStrengthEnum, default="Normal")
 
 jointType = {
     "SpineBase":0,
@@ -176,14 +202,13 @@ def initialize(context):
     context.scene.kmc_props.firstFramePosition = (-1,-1,-1)
     context.scene.kmc_props.initialOffset = (0,0,0)
 
-
     for target in context.scene.kmc_props.targetBones:
         if target.value is not None and target.value != "" :
             bone = bpy.data.objects[context.scene.kmc_props.arma_list].pose.bones[target.value]
             bone.rotation_mode = 'QUATERNION'
 
             # Store initial position of spine0 bone
-            if target.name == "Spine0":
+            if target.name == context.scene.kmc_props.rootBone:
                 context.scene.kmc_props.initialOffset = bpy.data.objects[context.scene.kmc_props.arma_list].pose.bones[target.value].matrix.translation
 
             # Store rest pose angles for column, head and feet bones
@@ -211,7 +236,7 @@ def updatePose(context, bone):
                 boneV = Vector((head[X] - tail[X], tail[Y] - head[Y], tail[Z] - head[Z]))
                 
                 # if first bone, update position (only for configured axes)
-                if target.name == "Spine0":
+                if target.name == context.scene.kmc_props.rootBone:
                     # initialize firstFramePosition if fit isn't
                     if context.scene.kmc_props.firstFramePosition[1] == -1:
                         context.scene.kmc_props.firstFramePosition = (-1.0*head[X], head[Y], head[Z])
@@ -245,7 +270,7 @@ def updatePose(context, bone):
                     context.scene.kmc_props.currentFrame += 1
                 if context.scene.kmc_props.record :
                     bone.keyframe_insert(data_path="rotation_quaternion", frame=context.scene.kmc_props.currentFrame)
-                    if target.name == "Spine0":
+                    if target.name == context.scene.kmc_props.rootBone:
                         bone.keyframe_insert(data_path="location", frame=context.scene.kmc_props.currentFrame)
                 
     # update child bones
@@ -290,6 +315,7 @@ class KinectMocapPanel(bpy.types.Panel):
                     if target.name == strBone :
                         box.prop(target, "value", text=target.name)
                         break
+            layout.prop(context.scene.kmc_props, "rootBone")
             
             # configure movement tracking
             layout.separator()
@@ -300,6 +326,10 @@ class KinectMocapPanel(bpy.types.Panel):
             col.prop(context.scene.kmc_props, "lockHeight")
             col.prop(context.scene.kmc_props, "lockwidth")
             
+            # denoising strength
+            layout.separator()
+            layout.prop(context.scene.kmc_props, "kalmanStrength")
+
             # activate
             layout.separator()
             layout.operator("kmc.start")
@@ -335,7 +365,16 @@ class KmcStartTrackingOperator(bpy.types.Operator):
     
     def modal(self, context, event):
         wm = context.window_manager
-        context.scene.k_sensor.init(1.0 / context.scene.kmc_props.fps)
+        uNoise = 5.0
+        if context.scene.kmc_props.kalmanStrength == "VeryLow" :
+            uNoise=50.0
+        elif context.scene.kmc_props.kalmanStrength == "Low" :
+            uNoise=20.0
+        elif context.scene.kmc_props.kalmanStrength == "Normal" :
+            uNoise=5.0
+        elif context.scene.kmc_props.kalmanStrength == "Strong" :
+            uNoise=1.0
+        context.scene.k_sensor.init(1.0 / context.scene.kmc_props.fps, 0.0005, uNoise)
         if event.type == 'TIMER':
             if(context.scene.k_sensor.update() == 1):
                 # update pose